Adding quiz questions are checked and made some fixes for the Ui Now quiz questions can be created which are properly following the schema for the questions although its seems a bit rough the pattern in which the questions are related to the quizes More Ui fix are made over the contributed quiz areas

+++ app/src/main/java/com/eva/firebasequizapp/contribute_quiz/presentation/CreateQuizEvents.kt
@@ -8,7 +8,7 @@ sealed class CreateQuizEvents {
     data class OnImageAdded(val uri: Uri? = null) : CreateQuizEvents()
 
     object OnImageRemoved : CreateQuizEvents()
-    data class OnColorAdded(val color: Int? = null) : CreateQuizEvents()
+    data class OnColorAdded(val color: ULong? = null) : CreateQuizEvents()
 
     object OnSubmit : CreateQuizEvents()
 }
